KKJJ-24 <refactor> (<User>) : 유저 경로변경 후 import 수정
경로 변경후 오류로 인해 바뀌지 않은  import 변경
