+++ src/main/java/renewal/awesome_travel/comment/entity/Comment.java
@@ -3,8 +3,8 @@ package renewal.awesome_travel.comment.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import renewal.awesome_travel.member.entity.User;
 import renewal.awesome_travel.product.entity.Product;
+import renewal.awesome_travel.user.entity.User;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
