+++ src/main/java/renewal/awesome_travel/comment/entity/CommentReport.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import renewal.awesome_travel.comment.utiles.ReportReason;
-import renewal.awesome_travel.member.entity.User;
+import renewal.awesome_travel.user.entity.User;
 
 import java.time.LocalDateTime;
