+++ src/main/java/renewal/awesome_travel/faq/entity/Faq.java
