+++ src/main/java/renewal/awesome_travel/inquiry/entity/Inquiry.java
@@ -2,7 +2,7 @@ package renewal.awesome_travel.inquiry.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
-import renewal.awesome_travel.member.entity.User;
+import renewal.awesome_travel.user.entity.User;
 
 import java.time.LocalDateTime;
