+++ src/main/java/renewal/awesome_travel/inquiry/repository/InquiryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import renewal.awesome_travel.inquiry.entity.Inquiry;
-import renewal.awesome_travel.member.entity.User;
+import renewal.awesome_travel.user.entity.User;
 
 public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
