+++ src/main/java/renewal/awesome_travel/inquiry/service/InquiryService.java
@@ -14,10 +14,10 @@ import renewal.awesome_travel.inquiry.entity.Inquiry;
 import renewal.awesome_travel.inquiry.entity.InquiryAnswer;
 import renewal.awesome_travel.inquiry.repository.InquiryAnswerRepository;
 import renewal.awesome_travel.inquiry.repository.InquiryRepository;
-import renewal.awesome_travel.member.entity.User;
-import renewal.awesome_travel.member.repository.UserRepository;
 import renewal.awesome_travel.notification.entity.Notification;
 import renewal.awesome_travel.notification.repository.NotificationRepository;
+import renewal.awesome_travel.user.entity.User;
+import renewal.awesome_travel.user.repository.UserRepository;
 
 @Service
 @RequiredArgsConstructor
