+++ src/main/java/renewal/awesome_travel/user/service/UserService.java
@@ -4,9 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import renewal.awesome_travel.member.utils.Provider;
-import renewal.awesome_travel.member.utils.Role;
-import renewal.awesome_travel.member.utils.Status;
 import renewal.awesome_travel.user.dto.request.PasswordChangeRequestDto;
 import renewal.awesome_travel.user.dto.request.UserRegisterRequestDto;
 import renewal.awesome_travel.user.dto.request.UserUpdateRequestDto;
@@ -16,6 +13,9 @@ import renewal.awesome_travel.user.entity.EmailVerificationToken;
 import renewal.awesome_travel.user.entity.User;
 import renewal.awesome_travel.user.repository.EmailVerificationTokenRepository;
 import renewal.awesome_travel.user.repository.UserRepository;
+import renewal.awesome_travel.user.utils.Provider;
+import renewal.awesome_travel.user.utils.Role;
+import renewal.awesome_travel.user.utils.Status;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
